+++ spec/app/domain/authentication/authn-oidc/v2/data_objects/authenticator.rb
@@ -69,4 +69,16 @@ RSpec.describe('Authentication::AuthnOidc::V2::DataObjects::Authenticator') do
       it { expect(authenticator.response_type).to eq('code') }
     end
   end
+
+  describe '.initialize' do
+    context 'when arguements are missing' do
+      let(:default_args) { { account: 'foo', service_id: 'baz' } }
+      it 'throws an exception' do
+        expect { authenticator }.to raise_error(
+          ArgumentError,
+          "missing keywords: :provider_uri, :client_id, :client_secret, :claim_mapping, :nonce, :state"
+        )
+      end
+    end
+  end
 end